Added qr code generation, moved graphing.
