+++ cogs/generation/graphing.py
