--- cogs/math/calculator.py
@@ -1,41 +1,0 @@
-"""
-> Xythrion
-> Copyright (c) 2020 Xithrius
-> MIT license, Refer to LICENSE for more info
-"""
-
-
-from discord.ext import commands as comms
-
-
-class Calculator(comms.Cog):
-    """Summary for Calculator
-
-    Attributes:
-        bot (:obj:`comms.Bot`): Represents a Discord bot.
-
-    """
-
-    def __init__(self, bot):
-        """Creating important attributes for this class.
-
-        Args:
-            bot (:obj:`comms.Bot`): Represents a Discord bot.
-
-        """
-        self.bot = bot
-
-    @comms.command()
-    async def multiply(self, ctx, *, eq: str):
-        """Multiplies two items together.
-
-        Args:
-            ctx (:obj:`comms.Context`): Represents the context in which a command is being invoked under.
-            eq (str): The equation to be parsed.
-
-        """
-        pass
-
-
-def setup(bot):
-    bot.add_cog(Calculator(bot))
